requirements.txt
python-dotenv
google-generativeai
python-docx
PyPDF2
PyYAML
streamlit
pandas
plotly
